# openapi-v1.yaml
openapi: 3.1.0
info:
  title: ARB AI Agent API
  description: Architecture Review Board AI Agent Backend
  version: 1.0.0
paths:
...
  /api/v1/artefacts/artefacts/upload:
    post:
      tags:
      - v1 / artefacts
      summary: Upload Artefact
      description: Upload an artefact for a review
      operationId: upload_artefact_api_v1_artefacts_artefacts_upload_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_artefact_api_v1_artefacts_artefacts_upload_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/artefacts/artefacts/review/{review_id}:
    get:
      tags:
      - v1 / artefacts
      summary: Get Review Artefacts
      description: Get all artefacts for a review
      operationId: get_review_artefacts_api_v1_artefacts_artefacts_review__review_id__get
      parameters:
      - name: review_id
        in: path
        required: true
        schema:
          type: string
          title: Review Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/artefacts/artefacts/chunks/{review_id}:
    get:
      tags:
      - v1 / artefacts
      summary: Get Review Chunks
      description: Get relevant chunks for a review
      operationId: get_review_chunks_api_v1_artefacts_artefacts_chunks__review_id__get
      parameters:
      - name: review_id
        in: path
        required: true
        schema:
          type: string
          title: Review Id
      - name: domain_slug
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Domain Slug
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/artefacts/artefacts/{artefact_id}:
    delete:
      tags:
      - v1 / artefacts
      summary: Delete Artefact
      description: Delete an artefact
      operationId: delete_artefact_api_v1_artefacts_artefacts__artefact_id__delete
      parameters:
      - name: artefact_id
        in: path
        required: true
        schema:
          type: string
          title: Artefact Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/artefacts/knowledge-base/search:
    get:
      tags:
      - v1 / artefacts
      summary: Search Knowledge Base
      description: Search knowledge base for relevant content
      operationId: search_knowledge_base_api_v1_artefacts_knowledge_base_search_get
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
          title: Query
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 5
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/artefacts/knowledge-base:
    post:
      tags:
      - v1 / artefacts
      summary: Create Knowledge Base Entry
      description: Create a knowledge base entry
      operationId: create_knowledge_base_entry_api_v1_artefacts_knowledge_base_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_create_knowledge_base_entry_api_v1_artefacts_knowledge_base_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - v1 / artefacts
      summary: Get Knowledge Base
      description: Get knowledge base entries
      operationId: get_knowledge_base_api_v1_artefacts_knowledge_base_get
      parameters:
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v2/artefacts/artefacts/upload:
    post:
      tags:
      - v2 / artefacts
      summary: Upload Artefact
      description: Upload an artefact for a review
      operationId: upload_artefact_api_v2_artefacts_artefacts_upload_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_artefact_api_v2_artefacts_artefacts_upload_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/artefacts/artefacts/review/{review_id}:
    get:
      tags:
      - v2 / artefacts
      summary: Get Review Artefacts
      description: Get all artefacts for a review
      operationId: get_review_artefacts_api_v2_artefacts_artefacts_review__review_id__get
      parameters:
      - name: review_id
        in: path
        required: true
        schema:
          type: string
          title: Review Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/artefacts/artefacts/chunks/{review_id}:
    get:
      tags:
      - v2 / artefacts
      summary: Get Review Chunks
      description: Get relevant chunks for a review
      operationId: get_review_chunks_api_v2_artefacts_artefacts_chunks__review_id__get
      parameters:
      - name: review_id
        in: path
        required: true
        schema:
          type: string
          title: Review Id
      - name: domain_slug
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Domain Slug
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/artefacts/artefacts/{artefact_id}:
    delete:
      tags:
      - v2 / artefacts
      summary: Delete Artefact
      description: Delete an artefact
      operationId: delete_artefact_api_v2_artefacts_artefacts__artefact_id__delete
      parameters:
      - name: artefact_id
        in: path
        required: true
        schema:
          type: string
          title: Artefact Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/artefacts/knowledge-base/search:
    get:
      tags:
      - v2 / artefacts
      summary: Search Knowledge Base
      description: Search knowledge base for relevant content
      operationId: search_knowledge_base_api_v2_artefacts_knowledge_base_search_get
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
          title: Query
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 5
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/artefacts/knowledge-base:
    post:
      tags:
      - v2 / artefacts
      summary: Create Knowledge Base Entry
      description: Create a knowledge base entry
      operationId: create_knowledge_base_entry_api_v2_artefacts_knowledge_base_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_create_knowledge_base_entry_api_v2_artefacts_knowledge_base_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - v2 / artefacts
      summary: Get Knowledge Base
      description: Get knowledge base entries
      operationId: get_knowledge_base_api_v2_artefacts_knowledge_base_get
      parameters:
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_create_knowledge_base_entry_api_v1_artefacts_knowledge_base_post:
      properties:
        title:
          type: string
          title: Title
        content:
          type: string
          title: Content
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        principle_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Principle Id
      type: object
      required:
      - title
      - content
      title: Body_create_knowledge_base_entry_api_v1_artefacts_knowledge_base_post
    Body_create_knowledge_base_entry_api_v2_artefacts_knowledge_base_post:
      properties:
        title:
          type: string
          title: Title
        content:
          type: string
          title: Content
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        principle_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Principle Id
      type: object
      required:
      - title
      - content
      title: Body_create_knowledge_base_entry_api_v2_artefacts_knowledge_base_post
...
    Body_upload_artefact_api_v1_artefacts_artefacts_upload_post:
      properties:
        review_id:
          type: string
          title: Review Id
        domain_slug:
          type: string
          title: Domain Slug
        artefact_name:
          type: string
          title: Artefact Name
        artefact_type:
          type: string
          title: Artefact Type
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - review_id
      - domain_slug
      - artefact_name
      - artefact_type
      - file
      title: Body_upload_artefact_api_v1_artefacts_artefacts_upload_post
...
    Body_upload_artefact_api_v2_artefacts_artefacts_upload_post:
      properties:
        review_id:
          type: string
          title: Review Id
        domain_slug:
          type: string
          title: Domain Slug
        artefact_name:
          type: string
          title: Artefact Name
        artefact_type:
          type: string
          title: Artefact Type
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - review_id
      - domain_slug
      - artefact_name
      - artefact_type
      - file
      title: Body_upload_artefact_api_v2_artefacts_artefacts_upload_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError